+++ TwitterNetworkLayerTests/NSURLCache+TNLAdditionsTest.m
@@ -75,10 +75,11 @@
     [[NSFileManager defaultManager] removeItemAtPath:path error:NULL];
     NSURLCache *cache = [[NSURLCache alloc] initWithMemoryCapacity:1024*1024*10 diskCapacity:1024*1024*10 diskPath:path];
     [cache removeAllCachedResponses];
-    [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:2.5]]; // give cache time to purge
+    [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:2.0]]; // give cache time to purge
     tnl_defer(^{
+        [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]]; // give cache time to store cache entries
         [cache removeAllCachedResponses];
-        [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:2.5]]; // give cache time to purge
+        [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:2.0]]; // give cache time to purge
     });
 
     TNLResponse *response = nil;
@@ -100,22 +101,9 @@
     XCTAssertEqualObjects(response.info.data, body);
     XCTAssertEqual(response.info.source, TNLResponseSourceNetworkRequest);
 
-#if TARGET_OS_OSX
-    // macOS has no way to clear the NSURLCache disk cache via API
-    // clearing the cache from disk can be done, but corrupts the sqlite db
-    // radar://43799833
-    return;
-#endif
-
-    if (tnl_available_ios_10) {
-        // ok
-    } else {
-        // just like macOS, iOS and tvOS on iOS 9 and lower cannot clear their disk cache
-        return;
-    }
-
+    [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]]; // give cache time to store cache entries
     [cache removeAllCachedResponses];
-    [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:2.5]]; // give cache time to purge
+    [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:2.0]]; // give cache time to purge
 
     config.cachePolicy = NSURLRequestReturnCacheDataDontLoad;
     response = [[self class] GETResponseWithURL:URL config:config];